backend/requirements.txt
fastapi
uvicorn
rembg[cpu]
pillow
numpy
opencv-python-headless
python-multipart
